+++ account-data-impl-db/src/se/liu/ida/tdp024/account/data/impl/db/facade/TransactionEntityFacadeDB.java
@@ -5,14 +5,19 @@
  */
 package se.liu.ida.tdp024.account.data.impl.db.facade;
 
+import se.liu.ida.tdp024.account.util.exceptions.ServiceConfigurationException;
+import se.liu.ida.tdp024.account.util.exceptions.InputParameterException;
+import se.liu.ida.tdp024.account.util.exceptions.EntityNotFoundException;
 import java.util.List;
 import javax.persistence.EntityManager;
+import javax.persistence.RollbackException;
 import javax.persistence.TypedQuery;
 import javax.persistence.criteria.CriteriaBuilder;
 import javax.persistence.criteria.CriteriaQuery;
 import javax.persistence.criteria.Root;
 import se.liu.ida.tdp024.account.data.api.entity.Transaction;
 import se.liu.ida.tdp024.account.data.api.facade.TransactionEntityFacade;
+import se.liu.ida.tdp024.account.data.api.util.*;
 import se.liu.ida.tdp024.account.data.impl.db.entity.TransactionDB;
 import se.liu.ida.tdp024.account.data.impl.db.util.EMF;
 import se.liu.ida.tdp024.account.util.logger.AccountLogger;
@@ -27,7 +32,9 @@ public class TransactionEntityFacadeDB implements TransactionEntityFacade {
     AccountLogger logger = new AccountLoggerImpl();
     
     @Override
-    public long create(String type, int amount, String status) {
+    public long create(String type, int amount, String status) throws 
+            InputParameterException, 
+            ServiceConfigurationException {
         
         EntityManager em = EMF.getEntityManager();
         
@@ -45,9 +52,12 @@ public class TransactionEntityFacadeDB implements TransactionEntityFacade {
             
             return transaction.getId();
         }
-        catch(Exception e) {
+        catch(IllegalArgumentException e) {
             logger.log(e);
-            return 0;
+            throw new InputParameterException("Wrong parameters.");
+        } catch(RollbackException e){
+            logger.log(e);
+            throw new ServiceConfigurationException("Server Error.");
         }
         finally {
             if(em.getTransaction().isActive()) {
@@ -58,22 +68,25 @@ public class TransactionEntityFacadeDB implements TransactionEntityFacade {
     }
 
     @Override
-    public Transaction find(long id) {
+    public Transaction find(long id) throws EntityNotFoundException, InputParameterException{
         EntityManager em = EMF.getEntityManager();
         
         try {
-            return em.find(TransactionDB.class, id);
-            
-        } catch (Exception e) {
+            Transaction result = em.find(TransactionDB.class, id);
+            if (result == null){
+                throw new EntityNotFoundException("The transaction does not exist.");
+            }
+            return result;
+        } catch (IllegalArgumentException e) {
             logger.log(e);
-            return null;
+            throw new InputParameterException("Wrong parameters.");
         } finally {
             em.close();
         }
     }
 
     @Override
-    public List<Transaction> findAll() {
+    public List<Transaction> findAll() throws ServiceConfigurationException{
         EntityManager em = EMF.getEntityManager();
         
         try {
@@ -90,9 +103,8 @@ public class TransactionEntityFacadeDB implements TransactionEntityFacade {
             List<Transaction> transactions = tq.getResultList();
             return transactions;
         } 
-        catch(Exception e) {
-            logger.log(e);
-            return null;
+        catch(IllegalArgumentException e) {
+            throw new ServiceConfigurationException("Server Error.");
         }
         finally {
             em.close();
@@ -100,11 +112,16 @@ public class TransactionEntityFacadeDB implements TransactionEntityFacade {
     }
 
     @Override
-    public void update(long id, String type, int amount, String status) throws Exception {
+    public void update(long id, String type, int amount, String status) throws 
+            EntityNotFoundException, 
+            ServiceConfigurationException {
         EntityManager em = EMF.getEntityManager();
         
         try {
             Transaction transaction = em.find(TransactionDB.class, id);
+            if (transaction == null){
+                throw new EntityNotFoundException("The transaction does not exist.");
+            }
             
             em.getTransaction().begin();
             transaction.setType(type);
@@ -112,9 +129,12 @@ public class TransactionEntityFacadeDB implements TransactionEntityFacade {
             transaction.setStatus(status);
             em.getTransaction().commit();
         }
-        catch (Exception e) {
+        catch (EntityNotFoundException e) {
             logger.log(e);
             throw e;
+        }
+        catch (RollbackException e){
+            throw new ServiceConfigurationException();
         }
         finally {
             if(em.getTransaction().isActive()) {
@@ -125,20 +145,25 @@ public class TransactionEntityFacadeDB implements TransactionEntityFacade {
     }
 
     @Override
-    public void remove(long id) throws Exception {
+    public void remove(long id) throws EntityNotFoundException, ServiceConfigurationException, InputParameterException {
         EntityManager em  = EMF.getEntityManager();
         
         
         try {
             Transaction transaction = em.find(TransactionDB.class, id);
+            if (transaction == null){
+                throw new EntityNotFoundException("The transaction does not exist.");
+            } 
             
             em.getTransaction().begin();
             em.remove(transaction);            
             em.getTransaction().commit();
-        }
-        catch (Exception e) {
+        } catch(IllegalArgumentException e){
             logger.log(e);
-            throw e;
+            throw new InputParameterException("Wrong parameters.");}
+        catch (RollbackException e) {
+            logger.log(e);
+            throw new ServiceConfigurationException("Server Error.");
         }
         finally {
             if(em.getTransaction().isActive()) {
